+++ astrality/tests/module/test_module_requires.py
@@ -1,15 +1,17 @@
+"""Tests for module requirements."""
+
 import logging
 from pathlib import Path
 
 from astrality.module import Module
+from astrality.tests.utils import RegexCompare
 
 def test_module_requires_option(caplog):
     """Test that modules are disabled when they don't satisfy `requires`."""
-
     # Simple module that satisfies the requirements
     does_satisfy_requiremnets = { 'module/satisfies': {
         'enabled': True,
-        'requires': 'command -v cd',
+        'requires': {'shell': 'command -v cd',},
     }}
     assert Module.valid_class_section(
         section=does_satisfy_requiremnets,
@@ -20,7 +22,7 @@ def test_module_requires_option(caplog):
 
     # Simple module that does not satisfy requirements
     does_not_satisfy_requirements = { 'module/does_not_satisfy': {
-        'requires': 'command -v does_not_exist',
+        'requires': {'shell': 'command -v does_not_exist',},
     }}
     assert not Module.valid_class_section(
         section=does_not_satisfy_requirements,
@@ -30,22 +32,29 @@ def test_module_requires_option(caplog):
     assert (
         'astrality',
         logging.WARNING,
-        '[module/does_not_satisfy] Module does not satisfy requirement "command -v does_not_exist".',
+        '[module/does_not_satisfy] Module requirements: '
+        'Unsuccessful command: "command -v does_not_exist", !',
     ) in caplog.record_tuples
 
 
     # Test failing one of many requirements
     does_not_satisfy_one_requirement = { 'module/does_not_satisfy': {
-        'requires': ['command -v cd', 'command -v does_not_exist'],
+        'requires': [
+            {'shell': 'command -v cd',},
+            {'shell': 'command -v does_not_exist',},
+        ],
     }}
     caplog.clear()
     assert not Module.valid_class_section(
-        section=does_not_satisfy_requirements,
+        section=does_not_satisfy_one_requirement,
         requires_timeout=1,
         requires_working_directory=Path('/'),
     )
     assert (
         'astrality',
         logging.WARNING,
-        '[module/does_not_satisfy] Module does not satisfy requirement "command -v does_not_exist".',
+        RegexCompare(
+            '\[module/does_not_satisfy\] Module requirements: .+ '
+            'Unsuccessful command: "command -v does_not_exist", !',
+        )
     ) in caplog.record_tuples